test(client): isolate visual asset loading contract
